proyecto_final/requirements.txt
asgiref==3.5.2
Django==3.2.16
djangorestframework==3.14.0
mysqlclient==2.1.0
pip==22.3.1
pytz==2022.6
setuptools==39.0.1
sqlparse==0.4.3
typing_extensions==4.4.0